Changed the skybox's layer to beneath base layer
Changed the skybox's layer to beneath the base layer. I did this so that graphical glitch where sometimes 3d boxes with transparent textures appear gray in the transparent bits because of it overlapping with the skybox doesn't happen. May need to add the option to turn off the skybox because it may cause headaches.